Move Record type into its own file
